move ingestion.py to staging app
